backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
openai
python-multipart
python-dotenv